ssl_verify_openssl: Avoid conversion warning in x509_verify_cert_ku
Just use the correct types.

v2:
 - Change type of expected_len argument to size_t

Acked-by: Gert Doering <gert@greenie.muc.de>
Gerrit URL: https://gerrit.openvpn.net/c/openvpn/+/1445
Message-Id: <20260119122058.14865-1-gert@greenie.muc.de>
URL: https://www.mail-archive.com/openvpn-devel@lists.sourceforge.net/msg35322.html

+++ src/openvpn/ssl_verify_backend.h
@@ -240,7 +240,7 @@ result_t x509_verify_ns_cert_type(openvpn_x509_cert_t *cert, const int usage);
  *                      if key usage is not enabled, or the values do not match.
  */
 result_t x509_verify_cert_ku(openvpn_x509_cert_t *x509, const unsigned *const expected_ku,
-                             int expected_len);
+                             size_t expected_len);
 
 /*
  * Verify X.509 extended key usage extension field.
